restructure project [pixfmt, color, tests/tutorial] (#10)
* split pixfmt into files

* cargo fmt

* split color.rs

* cargo fmt

* move buffer to pixfmt

* restructure examples

* fix format and clippy
