+++ images/t02_pixel_formats.png
