+++ src/color/rgb.rs
@@ -1,94 +1,11 @@
-//! Colors
-
 use palette::{
   blend::Premultiply,
   encoding::{FromLinear, IntoLinear},
 };
 
-use crate::{Color, ColorValue, U8, U16};
-
-pub trait FromColor {
-  fn from_color<C: Color>(c: C) -> Self;
-}
-
-pub trait FromRaw2: Sized {
-  type Raw;
-  fn from_raw(v1: Self::Raw, v2: Self::Raw) -> Self;
-  fn from_slice(slice: &[Self::Raw]) -> Self
-  where
-    Self::Raw: Copy,
-  {
-    assert!(slice.len() >= 2);
-    Self::from_raw(slice[0], slice[1])
-  }
-}
-pub trait IntoRaw2: Sized {
-  type Raw;
-  fn into_raw(self) -> (Self::Raw, Self::Raw);
-  fn into_slice(self) -> [Self::Raw; 2]
-  where
-    Self::Raw: Copy,
-  {
-    let (v1, v2) = self.into_raw();
-    [v1, v2]
-  }
-}
-
-pub trait FromRaw3: Sized {
-  type Raw;
-  fn from_raw(red: Self::Raw, green: Self::Raw, blue: Self::Raw) -> Self;
-  fn from_slice(slice: &[Self::Raw]) -> Self
-  where
-    Self::Raw: Copy,
-  {
-    assert!(slice.len() >= 3);
-    Self::from_raw(slice[0], slice[1], slice[2])
-  }
-}
-pub trait IntoRaw3: Sized {
-  type Raw;
-  fn into_raw(self) -> (Self::Raw, Self::Raw, Self::Raw);
-  fn into_slice(self) -> [Self::Raw; 3]
-  where
-    Self::Raw: Copy,
-  {
-    let (v1, v2, v3) = self.into_raw();
-    [v1, v2, v3]
-  }
-}
-
-pub trait FromRaw4: Sized {
-  type Raw;
-  fn from_raw(red: Self::Raw, green: Self::Raw, blue: Self::Raw, alpha: Self::Raw) -> Self;
-  fn from_slice(slice: &[Self::Raw]) -> Self
-  where
-    Self::Raw: Copy,
-  {
-    assert!(slice.len() >= 4);
-    Self::from_raw(slice[0], slice[1], slice[2], slice[3])
-  }
-}
-pub trait IntoRaw4: Sized {
-  type Raw;
-  fn into_raw(self) -> (Self::Raw, Self::Raw, Self::Raw, Self::Raw);
-  fn into_slice(self) -> [Self::Raw; 4]
-  where
-    Self::Raw: Copy,
-  {
-    let (v1, v2, v3, v4) = self.into_raw();
-    [v1, v2, v3, v4]
-  }
-}
-
-pub trait NamedColor {
-  const EMPTY: Self;
-  const WHITE: Self;
-  const BLACK: Self;
-  const RED: Self;
-  const GREEN: Self;
-  const BLUE: Self;
-  fn gray_color<T: ColorValue>(value: T) -> Self;
-}
+use crate::{
+  Color, ColorValue, FromColor, FromRaw2, FromRaw3, FromRaw4, IntoRaw2, IntoRaw3, IntoRaw4, NamedColor, U8, U16,
+};
 
 pub type Rgba<T> = palette::rgb::LinSrgba<T>;
 pub type Rgba8 = Rgba<U8>;
@@ -452,118 +369,5 @@ impl<T: ColorValue> Color for Gray<T> {
   }
   fn is_premultiplied(&self) -> bool {
     false
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-
-  #[test]
-  fn value_type_test() {
-    for i in 0..=255u8 {
-      assert_eq!(U8::new(i).as_color_::<U8>(), U8::new(i));
-    }
-    for i in 0..=65535u16 {
-      assert_eq!(U16::new(i).as_color_::<U16>(), U16::new(i));
-    }
-  }
-
-  #[test]
-  fn rgb8_to_gray8_test() {
-    let values = [
-      [0, 0, 0, 0u8],
-      [255, 255, 255, 255],
-      [255, 0, 0, 54],
-      [0, 255, 0, 182],
-      [0, 0, 255, 18],
-      [255, 255, 0, 237],
-      [255, 0, 255, 73],
-      [0, 255, 255, 201],
-      [128, 128, 128, 128],
-      [128, 0, 0, 27],
-      [0, 128, 0, 92],
-      [0, 0, 128, 9],
-      [128, 128, 0, 119],
-      [128, 0, 128, 36],
-      [0, 128, 128, 101],
-    ];
-    for [r, g, b, z] in &values {
-      let c = Rgb8::new(U8::new(*r), U8::new(*g), U8::new(*b));
-      let gray = c.gray8();
-      assert_eq!(gray.luma.0, *z);
-    }
-  }
-  #[test]
-  fn rgb8_test() {
-    let w = Rgb8::WHITE;
-    assert_eq!(w.into_raw(), (255, 255, 255));
-    assert_eq!((w.red8(), w.green8(), w.blue8()), (255, 255, 255));
-    let w = Rgb8::BLACK;
-    assert_eq!(w.into_raw(), (0, 0, 0));
-    assert_eq!((w.red8(), w.green8(), w.blue8()), (0, 0, 0));
-    let w = Rgb8::gray_color(0.5);
-    assert_eq!(w.into_raw(), (128, 128, 128));
-    assert_eq!((w.red8(), w.green8(), w.blue8()), (128, 128, 128));
-    // let w = Rgb8::from_slice(&[1, 2, 3]);
-    // assert_eq!(w, Rgb8::from_raw(1, 2, 3));
-    let w = Rgb8::from_raw(0, 90, 180);
-    assert_eq!(w.into_raw(), (0, 90, 180));
-    assert_eq!((w.red8(), w.green8(), w.blue8()), (0, 90, 180));
-  }
-  #[test]
-  fn gray_test() {
-    let g = Gray8::from_raw(34, 255);
-    assert_eq!(g.into_raw(), (34, 255));
-    let g = Gray8::from_raw(134, 100);
-    assert_eq!(g.into_raw(), (134, 100));
-    // let g = Gray8::from_slice(&[10,20]);
-    // assert_eq!(g.into_raw(), (10,20));
-  }
-  #[test]
-  fn rgba8_test() {
-    let c = Rgba8::WHITE;
-    assert_eq!(c.into_raw(), (255, 255, 255, 255));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (255, 255, 255, 255));
-    let c = Rgba8::BLACK;
-    assert_eq!(c.into_raw(), (0, 0, 0, 255));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (0, 0, 0, 255));
-    let c = Rgba8::from_raw(255, 90, 84, 72);
-    assert_eq!(c.into_raw(), (255, 90, 84, 72));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (255, 90, 84, 72));
-    let c = Rgba8::EMPTY;
-    assert_eq!(c.into_raw(), (0, 0, 0, 0));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (0, 0, 0, 0));
-    let c = Rgba8::from_raw(255, 255, 255, 128);
-    let p = c.premultiply();
-    assert_eq!(p, RgbaPre8::from_raw(128, 128, 128, 128));
-  }
-  #[test]
-  fn srgb_test() {
-    let s = Srgba8::from_raw(50, 150, 250, 128);
-    assert_eq!(s.into_raw(), (50, 150, 250, 128));
-    assert_eq!((s.red8(), s.green8(), s.blue8(), s.alpha8()), (8, 78, 244, 128));
-    let t = s.rgba8();
-    assert_eq!(t.into_raw(), (8, 78, 244, 128));
-    assert_eq!((t.red8(), t.green8(), t.blue8(), t.alpha8()), (8, 78, 244, 128));
-    let s2 = t.srgba8();
-    assert_eq!(s2.into_raw(), (50, 150, 250, 128));
-    assert_eq!((s2.red8(), s2.green8(), s2.blue8(), s2.alpha8()), (8, 78, 244, 128));
-  }
-
-  #[test]
-  fn rgba_pre_test() {
-    let c = RgbaPre8::WHITE;
-    assert_eq!(c.into_raw(), (255, 255, 255, 255));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (255, 255, 255, 255));
-    let c = RgbaPre8::BLACK;
-    assert_eq!(c.into_raw(), (0, 0, 0, 255));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (0, 0, 0, 255));
-    let c = RgbaPre8::from_raw(255, 90, 84, 72);
-    assert_eq!(c.into_raw(), (255, 90, 84, 72));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (255, 90, 84, 72));
-    let c = Rgba8::from_raw(255, 90, 84, 72).premultiply();
-    assert_eq!(c, RgbaPre8::from_raw(72, 25, 24, 72));
-    assert_eq!((c.red8(), c.green8(), c.blue8(), c.alpha8()), (72, 25, 24, 72));
   }
 }
