+++ src/outline.rs
@@ -2,7 +2,7 @@
 //!
 //! ```
 //! use agg::prelude::*;
-//! let pix = Pixfmt::<Rgb8>::new(100, 100);
+//! let pix = Pixfmt::<Rgb8>::create(100, 100);
 //! let mut ren_base = agg::RenderingBase::new(pix);
 //! ren_base.clear(Rgb8::WHITE);
 //!
