+++ tests/component_rendering.rs
@@ -4,7 +4,7 @@ use agg::prelude::*;
 fn draw(alpha: u8, filename: &str) {
   let (w, h) = (320, 320);
 
-  let pixf = agg::Pixfmt::<Rgb8>::new(w, h);
+  let pixf = agg::Pixfmt::<Rgb8>::create(w, h);
   let mut ren_base = agg::RenderingBase::new(pixf);
   ren_base.clear(Rgb8::WHITE);
   let g8 = Gray8::from_raw(0, alpha);
