+++ tests/test_tutorials.rs
@@ -1,0 +1,1 @@
+mod tutorials;
