+++ tests/tutorials/mod.rs
@@ -1,0 +1,17 @@
+// https://agg.sourceforge.net/antigrain.com/doc/basic_renderers/basic_renderers.agdoc.html
+
+pub mod t00_example;
+pub mod t01_rendering_buffer;
+pub mod t02_pixel_formats;
+pub mod t03_solar_spectrum;
+pub mod t04_solar_spectrum_alpha;
+pub mod t05_solar_spectrum_alpha;
+pub mod t11;
+pub mod t12;
+pub mod t13;
+pub mod t14;
+pub mod t15;
+pub mod t16;
+pub mod t21_line_join;
+pub mod t22_inner_join;
+pub mod t23_font;
