+++ tests/tutorials/t01_rendering_buffer.rs
@@ -17,7 +17,7 @@ fn draw_black_frame(pix: &mut agg::Pixfmt<agg::Rgb8>) {
 #[test]
 fn t01_rendering_buffer() {
   //let mut rbuf = RenderingBuffer::new(320, 220, 3);
-  let mut pix = agg::Pixfmt::<agg::Rgb8>::new(320, 220);
+  let mut pix = agg::Pixfmt::<agg::Rgb8>::create(320, 220);
   for i in 0..pix.width() {
     for j in 0..pix.height() {
       pix.set((i, j), agg::Rgb8::WHITE);
@@ -30,6 +30,6 @@ fn t01_rendering_buffer() {
     pix.set((i, i), agg::Rgb8::from_raw(127, 200, 98));
   }
 
-  pix.to_file("tests/tmp/agg_test_01.png").unwrap();
-  assert!(agg::ppm::img_diff("tests/tmp/agg_test_01.png", "images/agg_test_01.png").unwrap());
+  pix.to_file("tests/tmp/t01_rendering_buffer.png").unwrap();
+  assert!(agg::ppm::img_diff("tests/tmp/t01_rendering_buffer.png", "images/t01_rendering_buffer.png").unwrap());
 }
