+++ tests/tutorials/t11.rs
@@ -1,12 +1,11 @@
 extern crate agg;
-
-use agg::prelude::*;
+use agg::{color::NamedColor, prelude::*};
 
 #[test]
-fn t14_with_gamma() {
+fn t11_full() {
   let (w, h) = (100, 100);
 
-  let pixf = agg::Pixfmt::<agg::Rgb8>::new(w, h);
+  let pixf = agg::Pixfmt::<agg::Rgb8>::create(w, h);
 
   let mut ren_base = agg::RenderingBase::new(pixf);
 
@@ -16,15 +15,13 @@ fn t14_with_gamma() {
 
   let mut ras = agg::RasterizerScanline::new();
 
-  ras.clip_box(40.0, 0.0, w as f64 - 40.0, h as f64);
-
   ras.move_to(10.0, 10.0);
   ras.line_to(50.0, 90.0);
   ras.line_to(90.0, 10.0);
 
   agg::render_scanlines(&mut ras, &mut ren);
 
-  ren.to_file("tests/tmp/agg_test_14.png").unwrap();
+  ren.to_file("tests/tmp/agg_test_11.png").unwrap();
 
-  assert!(agg::ppm::img_diff("tests/tmp/agg_test_14.png", "images/agg_test_14.png").unwrap());
+  assert!(agg::ppm::img_diff("tests/tmp/agg_test_11.png", "images/agg_test_11.png").unwrap());
 }
